+++ src/Release.php
@@ -16,7 +16,7 @@ class Release
     /**
      * Version of the SDK
      */
-    const VERSION = '1.7.179';
+    const VERSION = '1.7.180';
 
     /**
      * @param Event $event